style: Rename files, add `make dataset`
